fixing mistake in endpoints file, changing source and target of the JAR file to 17 instead of 1.8, fixing the item class variables datatype, create ItemDTO class to work with controller, changing the code in most of classes that refere to createNewItem function (endpoint)

+++ Source/src/main/java/org/bringme/repository/ItemRepository.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface ItemRepository {
     List<Item> getAll();
     Optional<Item> getById(Long id);
-    int saveItem(Item item);
+    Long saveItem(Item item);
 }
